+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- request method to support starksign requests
 
 ## [0.0.1] - 2023-01-26
 ### Added